Fix/Refactor: kakao load 함수 util 로 분리 및 기능 오류 수정 카카오 공유 동작하지 않는 문제 수정
